Reduce number of metrics reporting threads
