--- memq-client/src/main/java/com/pinterest/memq/client/producer/http/DaemonThreadFactory.java
@@ -1,38 +1,0 @@
-/**
- * Copyright 2022 Pinterest, Inc.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *    http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.pinterest.memq.client.producer.http;
-
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicInteger;
-
-public final class DaemonThreadFactory implements ThreadFactory {
-
-  private String basename;
-  private AtomicInteger counter;
-
-  public DaemonThreadFactory(String basename) {
-    this.basename = basename;
-    this.counter = new AtomicInteger();
-  }
-
-  @Override
-  public Thread newThread(Runnable r) {
-    Thread th = new Thread(r);
-    th.setDaemon(true);
-    th.setName(basename + "-" + counter.getAndIncrement());
-    return th;
-  }
-}
